requirements.txt
pillow==8.1.2
pytesseract==0.3.7

